Prevent stack buffer overflow in ox sax parse (#396)
* Fixes shift type in ox_sax_buf_read

The shift variable can't be negative of size_t is used. It prevents realloc
logic to work as expected.

* Include SAX tests in test_all rake task

Remove OptionParser from sax_tests.rb to allow normal usage

* Improve code style to respect clang format

+++ ext/ox/sax_buf.c
@@ -69,8 +69,8 @@ void ox_sax_buf_init(Buf buf, VALUE io) {
 }
 
 int ox_sax_buf_read(Buf buf) {
-    int    err;
-    size_t shift = 0;
+    int  err;
+    long shift = 0;
 
     // if there is not much room to read into, shift or realloc a larger buffer.
     if (buf->head < buf->tail && 4096 > buf->end - buf->tail) {
